app/requirements.txt
flask>=1.0.2
mysql-connector>= 2.1.6
python-dotenv>=0.10.1
